refactor(session): isolate trace construction into a mixin
Four methods on ``SessionRunUseCase`` only ever read session state and wrote
the tracer: the ``context_shaping`` report around shaping, the
``steering_nudge`` high-water crossing, the ``llm_call`` row, and the
``session_terminal`` closing row. Three lived in the completion helpers and
one in the run loop, so the trajectory a run leaves behind had no single
place to be read from.

They now sit together in ``_session_run_trace.py`` as
``_SessionRunTraceMixin``, mixed into ``SessionRunUseCase`` beside the
completion mixin. Signatures, bodies, and the ``self`` attributes they touch
are unchanged; the two private marks they advance (``_last_steering_level``,
``_session_terminal_traced``) stay owned by the host's ``__init__`` and
``reset_for_restore``. The mixin names the host attributes it reads with bare
annotations, which create nothing at runtime and let pyright check the module
on its own. The ``commit_brake`` row stays with the enforcement gate that
trips it.

The ``tool_choice`` rejection classifier, a pure function of the provider
error, moves from the completion helpers to ``_session_run_shared`` next to
``_submit_tool_choice``; the completion module imports it from there.

One observable difference: the "session terminal trace failed" log line is
now emitted under this module's logger name rather than ``session_run``'s.
